# generate/server/go/go/model_create_or_update_todolist.go
/*
 * TodoList RESTful API
 *
 * OpenAPI for TodoList RESTful API
 *
 * API version: 1
 * Contact: yogadimaspambudi@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CreateOrUpdateTodolist struct {

	Name string `json:"name"`

	Priority float64 `json:"priority"`

	Tags []string `json:"tags,omitempty"`
}