# generate/server/kotlin/src/main/kotlin/io/swagger/server/models/ArrayTodolist.kt
/**
* TodoList RESTful API
* OpenAPI for TodoList RESTful API
*
* OpenAPI spec version: 1
* Contact: yogadimaspambudi@gmail.com
*
* NOTE: This class is auto generated by the swagger code generator program.
* https://github.com/swagger-api/swagger-codegen.git
* Do not edit the class manually.
*/package io.swagger.server.models

import io.swagger.server.models.Todolist

/**
 * 
 */
class ArrayTodolist (

) {
}