# generate/server/kotlin/src/main/kotlin/io/swagger/server/models/Todolist.kt
/**
* TodoList RESTful API
* OpenAPI for TodoList RESTful API
*
* OpenAPI spec version: 1
* Contact: yogadimaspambudi@gmail.com
*
* NOTE: This class is auto generated by the swagger code generator program.
* https://github.com/swagger-api/swagger-codegen.git
* Do not edit the class manually.
*/package io.swagger.server.models


/**
 * 
 * @param id 
 * @param name 
 * @param priority 
 * @param tags 
 */
data class Todolist (

    val id: kotlin.String? = null,
    val name: kotlin.String? = null,
    val priority: java.math.BigDecimal? = null,
    val tags: kotlin.Array<kotlin.String>? = null
) {
}