# generate/server/nodejs/api/openapi.yaml
openapi: 3.0.3
info:
  title: TodoList RESTful API
  description: OpenAPI for TodoList RESTful API
  termsOfService: https://github.com/Yogadimas
  contact:
    name: Yoga Dimas Pambudi
    url: https://github.com/Yogadimas
    email: yogadimaspambudi@gmail.com
  license:
    name: APACHE 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1"
externalDocs:
  description: Github Yoga Dimas Pambudi
  url: https://github.com/Yogadimas
servers:
- url: "https://{environment}.yogadimaspambudi.com/api/v1"
  description: TodoList RESTful API Server
  variables:
    environment:
      description: Server 1 Environment
      default: dev
      enum:
      - dev
      - qa
      - prod
- url: "https://{environment2}.yogadimaspambudi.com/api/v2"
  variables:
    environment2:
      default: dev
      enum:
      - dev
      - prod
- url: "https://{environment}.yogadimaspambudi.com/api/v3"
  variables:
    environment:
      default: dev
      enum:
      - dev
      - test
      - prod
paths:
  /todolist:
    get:
      tags:
      - All
      - Todolist
      summary: Get All Todolist
      description: Get all active todolist by default
      operationId: todolistGET
      parameters:
      - name: include_done
        in: query
        description: Is include_done todolist
        required: false
        style: form
        explode: true
        schema:
          type: boolean
          nullable: true
          default: false
      - name: name
        in: query
        description: Filter todolist by name
        required: false
        style: form
        explode: true
        schema:
          maxLength: 100
          minLength: 1
          type: string
          nullable: true
        examples:
          java:
            description: Example name Java
            value: Java
          php:
            description: Example name PHP
            value: PHP
      responses:
        "200":
          description: Success get all todolist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayTodolist'
              examples:
                success:
                  description: Example success get all todolist
                  value:
                  - id: "1"
                    name: Java
                    priority: 3
                    tags:
                    - Java
                    - Programming
                  - id: "2"
                    name: PHP
                    priority: 5
                    tags:
                    - PHP
                    - Programming
      security:
      - TodolistAuth: []
      x-swagger-router-controller: All
    post:
      tags:
      - Todolist
      summary: Create new Todolist
      description: Create new active todolist
      operationId: todolistPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateTodolist'
            examples:
              java:
                description: Example Create todolist Java
                value:
                  name: Java
                  priority: 3
                  tags:
                  - Java
                  - Programming
              php:
                description: Example Create todolist PHP
                value:
                  name: PHP
                  priority: 5
                  tags:
                  - PHP
                  - Programming
        required: true
      responses:
        "200":
          description: Success create todolist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todolist'
              examples:
                java:
                  description: Success create Java todolist
                  value:
                    id: "1"
                    name: Java
                    priority: 3
                    tags:
                    - Java
                    - Programming
                php:
                  description: Success create PHP todolist
                  value:
                    id: "2"
                    name: PHP
                    priority: 5
                    tags:
                    - PHP
                    - Programming
      security:
      - TodolistAuth: []
      x-swagger-router-controller: Todolist
  /todolist/{todolistId}:
    put:
      tags:
      - Todolist
      summary: Update existing todolist
      description: Update existing todolist in database
      operationId: todolistTodolistIdPUT
      parameters:
      - name: todolistId
        in: path
        description: TodoList id for updated
        required: true
        style: simple
        explode: false
        schema:
          maxLength: 100
          minLength: 1
          type: string
        examples:
          java:
            description: Sample todolist id for Java
            value: "1"
          php:
            description: Sample todolist id for PHP
            value: "2"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateTodolist'
            examples:
              java:
                description: Example Create todolist Java
                value:
                  name: Java
                  priority: 3
                  tags:
                  - Java
                  - Programming
              php:
                description: Example Create todolist PHP
                value:
                  name: PHP
                  priority: 5
                  tags:
                  - PHP
                  - Programming
        required: true
      responses:
        "200":
          description: Success update todolist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todolist'
              examples:
                java:
                  description: Success update Java todolist
                  value:
                    id: "1"
                    name: Java
                    priority: 3
                    tags:
                    - Java
                    - Programming
                php:
                  description: Success update PHP todolist
                  value:
                    id: "2"
                    name: PHP
                    priority: 5
                    tags:
                    - PHP
                    - Programming
      security:
      - TodolistAuth: []
      x-swagger-router-controller: Todolist
    delete:
      tags:
      - Todolist
      summary: Delete existing todolist
      description: Delete existing todolist in database
      operationId: todolistTodolistIdDELETE
      parameters:
      - name: todolistId
        in: path
        description: TodoList id for updated
        required: true
        style: simple
        explode: false
        schema:
          maxLength: 100
          minLength: 1
          type: string
        examples:
          java:
            description: Sample todolist id for Java
            value: "1"
          php:
            description: Sample todolist id for PHP
            value: "2"
      responses:
        "200":
          description: Success delete todolist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
              examples:
                success:
                  description: Success delete todolist
                  value:
                    success: true
        "404":
          description: Failed delete not found todolist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
              examples:
                failed:
                  description: Failed delete todolist
                  value:
                    success: false
      security:
      - TodolistAuth: []
      x-swagger-router-controller: Todolist
components:
  schemas:
    Todolist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        priority:
          type: number
          format: int32
        tags:
          type: array
          items:
            type: string
      example:
        name: name
        id: id
        priority: 0.80082819046101150206595775671303272247314453125
        tags:
        - tags
        - tags
    CreateOrUpdateTodolist:
      required:
      - name
      - priority
      type: object
      properties:
        name:
          type: string
        priority:
          type: number
          format: int32
        tags:
          type: array
          items:
            type: string
    ArrayTodolist:
      type: array
      items:
        $ref: '#/components/schemas/Todolist'
    inline_response_200:
      type: object
      properties:
        success:
          type: boolean
      example:
        success: true
  parameters:
    TodolistId:
      name: todolistId
      in: path
      description: TodoList id for updated
      required: true
      style: simple
      explode: false
      schema:
        maxLength: 100
        minLength: 1
        type: string
      examples:
        java:
          description: Sample todolist id for Java
          value: "1"
        php:
          description: Sample todolist id for PHP
          value: "2"
  securitySchemes:
    TodolistAuth:
      type: apiKey
      description: Authentication for Todolist RESTful API
      name: X-API-Key
      in: header